+++ MinimedKit/PumpEvents/BolusWizardEstimatePumpEvent.swift
@@ -13,14 +13,14 @@ public class BolusWizardEstimatePumpEvent: TimestampedPumpEvent {
   let timestamp: NSDateComponents
   let carbohydrates: Int
   let bloodGlucose: Int
-  let foodEstimate: Float
-  let correctionEstimate: Float
-  let bolusEstimate: Float
-  let unabsorbedInsulinTotal: Float
+  let foodEstimate: Double
+  let correctionEstimate: Double
+  let bolusEstimate: Double
+  let unabsorbedInsulinTotal: Double
   let bgTargetLow: Int
   let bgTargetHigh: Int
   let insulinSensitivity: Int
-  let carbRatio: Float
+  let carbRatio: Double
 
   public required init?(availableData: NSData, pumpModel: PumpModel) {
     
@@ -28,8 +28,8 @@ public class BolusWizardEstimatePumpEvent: TimestampedPumpEvent {
       return Int(availableData[idx] as UInt8)
     }
     
-    func insulinDecode(a: Int, b: Int) -> Float {
-      return Float((a << 8) + b) / 40.0
+    func insulinDecode(a: Int, b: Int) -> Double {
+      return Double((a << 8) + b) / 40.0
     }
     
     if pumpModel.larger {
@@ -59,24 +59,24 @@ public class BolusWizardEstimatePumpEvent: TimestampedPumpEvent {
       carbohydrates = ((d(8) & 0xc) << 6) + d(7)
       bloodGlucose = ((d(8) & 0x3) << 8) + d(1)
       foodEstimate = insulinDecode(d(14), b: d(15))
-      correctionEstimate = Float(((d(16) & 0b111000) << 5) + d(13)) / 40.0
+      correctionEstimate = Double(((d(16) & 0b111000) << 5) + d(13)) / 40.0
       bolusEstimate = insulinDecode(d(19), b: d(20))
       unabsorbedInsulinTotal = insulinDecode(d(17), b: d(18))
       bgTargetLow = d(12)
       bgTargetHigh = d(21)
       insulinSensitivity = d(11)
-      carbRatio = Float(((d(9) & 0x7) << 8) + d(10)) / 10.0
+      carbRatio = Double(((d(9) & 0x7) << 8) + d(10)) / 10.0
     } else {
       carbohydrates = d(7)
       bloodGlucose = ((d(8) & 0x3) << 8) + d(1)
-      foodEstimate = Float(d(13))/10.0
-      correctionEstimate = Float((d(14) << 8) + d(12)) / 10.0
-      bolusEstimate = Float(d(18))/10.0
-      unabsorbedInsulinTotal = Float(d(16))/10.0
+      foodEstimate = Double(d(13))/10.0
+      correctionEstimate = Double((d(14) << 8) + d(12)) / 10.0
+      bolusEstimate = Double(d(18))/10.0
+      unabsorbedInsulinTotal = Double(d(16))/10.0
       bgTargetLow = d(11)
       bgTargetHigh = d(19)
       insulinSensitivity = d(10)
-      carbRatio = Float(d(9))
+      carbRatio = Double(d(9))
     }
 
   }
